Restructure FE directory
jsx files for pages in the application will be found in the "pages" folder
jsx files for ui components will be found in the "components" folder

- components aimed to be as reusable as possible
