+++ src/components/Topbar/Topbar.jsx
@@ -1,9 +1,9 @@
 import React from 'react';
 import { useNavigate } from 'react-router';
-import { useAuth } from '../../../contexts/AuthContext';
-import { Sidebar } from '../Sidebar';
+import { useAuth } from '../../contexts/AuthContext';
+import { Sidebar } from '../Sidebar/Sidebar';
 import { StyledTopbar, NavbarText } from './StyledTopbar';
-import { FlexBoxCenter } from '../../styled/general';
+import { FlexBoxCenter } from '../styled/general';
 import { Box } from '@mui/material';
 
 export const Topbar = () => {
